requirements.txt
numpy
matplotlib
pandas
sklearn
tensorflow
keras==2.1.5
pyrebase
flask
gunicorn
pillow
